# server/src/public/swagger.json
{
      "openapi": "3.0.0",
      "info": {
            "title": "MyGame-V1",
            "description": "a board game build base on NestJs and NextJs",
            "contact": {
                  "name": "pham vinh nhan",
                  "email": "heaty566@gmail.com"
            },
            "license": {
                  "name": "Apache 2.0",
                  "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
            },
            "version": "0.0.1"
      },
...
      "paths": {
...
            "/api/user": {
                  "get": {
                        "tags": ["user"],
                        "parameters": [{ "$ref": "#/components/parameters/token" }],
                        "responses": {
                              "200": {
                                    "description": "return user information without some sensitive fields",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },

                              "401": {
                                    "$ref": "#/components/responses/401"
                              }
                        }
                  },
                  "put": {
                        "tags": ["user"],
                        "description": "update basic user information",
                        "parameters": [{ "$ref": "#/components/parameters/token" }],
                        "requestBody": {
                              "$ref": "#/components/requestBodies/UpdateBasicUserDto"
                        },
                        "responses": {
                              "200": {
                                    "description": "user update successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              },
                              "400": {
                                    "$ref": "#/components/responses/400"
                              }
                        }
                  }
            },
            "/api/user/avatar": {
                  "put": {
                        "tags": ["user"],
                        "description": "upload user avatar image",
                        "parameters": [{ "$ref": "#/components/parameters/token" }],
                        "requestBody": {
                              "$ref": "#/components/requestBodies/UpdateAvatarDto"
                        },
                        "responses": {
                              "200": {
                                    "description": "user update successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              },
                              "400": {
                                    "$ref": "#/components/responses/400"
                              },
                              "500": {
                                    "$ref": "#/components/responses/500"
                              }
                        }
                  }
            },
            "/api/user/password/:otp": {
                  "put": {
                        "tags": ["user"],
                        "description": "reset password with otp key",
                        "parameters": [{ "$ref": "#/components/parameters/token" }, { "$ref": "#/components/parameters/OTPkey" }],
                        "responses": {
                              "200": {
                                    "description": "update user successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "400": {
                                    "$ref": "#/components/responses/400"
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              }
                        }
                  }
            },
            "/api/user/email/:otp": {
                  "put": {
                        "tags": ["user"],
                        "description": "update email with otp key",
                        "parameters": [{ "$ref": "#/components/parameters/token" }, { "$ref": "#/components/parameters/OTPkey" }],
                        "responses": {
                              "200": {
                                    "description": "update user successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "400": {
                                    "$ref": "#/components/responses/400"
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              }
                        }
                  }
            },
            "/api/user/phone/:otp": {
                  "put": {
                        "tags": ["user"],
                        "description": "update phone with otp key",
                        "parameters": [{ "$ref": "#/components/parameters/token" }, { "$ref": "#/components/parameters/OTPkey" }],
                        "responses": {
                              "200": {
                                    "description": "update user successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "400": {
                                    "$ref": "#/components/responses/400"
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              }
                        }
                  }
            },
            "/api/user/otp-sms": {
                  "post": {
                        "tags": ["user"],
                        "description": "create new otp key and send to phone number",
                        "parameters": [{ "$ref": "#/components/parameters/token" }],
                        "requestBody": {
                              "$ref": "#/components/requestBodies/OtpSmsDTO"
                        },
                        "responses": {
                              "200": {
                                    "description": "update user successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "400": {
                                    "$ref": "#/components/responses/400"
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              },
                              "500": {
                                    "$ref": "#/components/responses/500"
                              }
                        }
                  }
            },
            "/api/user/otp-email": {
                  "post": {
                        "tags": ["user"],
                        "description": "create new otp key and send to email",
                        "parameters": [{ "$ref": "#/components/parameters/token" }],
                        "requestBody": {
                              "$ref": "#/components/requestBodies/UpdateEmailDto"
                        },
                        "responses": {
                              "200": {
                                    "description": "update user successfully",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "400": {
                                    "$ref": "#/components/responses/400"
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              },
                              "500": {
                                    "$ref": "#/components/responses/500"
                              }
                        }
                  }
            },
            "/api/admin/users": {
                  "get": {
                        "tags": ["admin"],
                        "description": "get all users information except some sensitive information",
                        "parameters": [{ "$ref": "#/components/parameters/token" }],
                        "responses": {
                              "200": {
                                    "description": "return array of user information",
                                    "content": {
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/CommonApiResponse"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "$ref": "#/components/responses/401"
                              },
                              "403": {
                                    "$ref": "#/components/responses/403"
                              }
                        }
                  }
            },
...
      },
      "components": {
            "responses": {
                  "401": {
                        "description": "user has not a re-token",
                        "content": {
                              "application/json": {
                                    "schema": {
                                          "$ref": "#/components/schemas/CommonApiResponse"
                                    }
                              }
                        }
                  },
                  "400": {
                        "description": "invalid body input",
                        "content": {
                              "application/json": {
                                    "schema": {
                                          "$ref": "#/components/schemas/CommonApiResponse"
                                    }
                              }
                        }
                  },
                  "500": {
                        "description": "some went wrong( 3rd shutdown or server errors)",
                        "content": {
                              "application/json": {
                                    "schema": {
                                          "$ref": "#/components/schemas/CommonApiResponse"
                                    }
                              }
                        }
                  },
                  "403": {
                        "description": "user has to be an administrator",
                        "content": {
                              "application/json": {
                                    "schema": {
                                          "$ref": "#/components/schemas/CommonApiResponse"
                                    }
                              }
                        }
                  },
...
            },
            "schemas": {
                  "CommonApiResponse": {
                        "type": "object",
                        "properties": {
                              "data": {
                                    "type": "object",
                                    "description": "it type any base on your api"
                              },
                              "message": {
                                    "type": "string",
                                    "example": "this is an error message"
                              },
                              "details": {
                                    "type": "object",
                                    "properties": {
                                          "field": {
                                                "type": "string",
                                                "description": "field will be the name of invalid field",
                                                "example": "should not be empty"
                                          }
                                    }
                              }
                        }
                  },
...
            },
            "requestBodies": {
                  "UpdateBasicUserDto": {
                        "content": {
                              "application/json": {
                                    "schema": {
                                          "type": "object",
                                          "properties": {
                                                "name": {
                                                      "type": "string",
                                                      "example": "who am i"
                                                }
                                          }
                                    }
                              }
                        },
                        "description": "user basic information is not include some sensitive fields",
                        "required": true
                  },
                  "UpdateAvatarDto": {
                        "description": "put avatar image here",
                        "content": {
                              "form-data": {
                                    "schema": {
                                          "type": "object",
                                          "properties": {
                                                "avatar": {
                                                      "type": "string",
                                                      "example": "/hello.png"
                                                }
                                          }
                                    }
                              }
                        },
                        "required": true
                  },
                  "OtpSmsDTO": {
                        "description": "provide phone number to send an otp key",
                        "content": {
                              "form-data": {
                                    "schema": {
                                          "type": "object",
                                          "properties": {
                                                "phoneNumber": {
                                                      "type": "string",
                                                      "example": "+20401230312"
                                                }
                                          }
                                    }
                              }
                        },
                        "required": true
                  },
                  "UpdateEmailDto": {
                        "description": "provide an email to send an otp key",
                        "content": {
                              "form-data": {
                                    "schema": {
                                          "type": "object",
                                          "properties": {
                                                "phoneNumber": {
                                                      "type": "string",
                                                      "example": "hello@gmail.com"
                                                }
                                          }
                                    }
                              }
                        },
                        "required": true
                  },
...
            },
            "parameters": {
                  "token": {
                        "name": "token",
                        "in": "cookie",
                        "required": true,
                        "style": "form",
                        "schema": {
                              "type": "object",
                              "properties": {
                                    "re-token": {
                                          "type": "string",
                                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                    },
                                    "auth-token": {
                                          "type": "string",
                                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                    }
                              }
                        }
                  },
...
                  "OTPkey": {
                        "name": "id",
                        "description": "otp key",
                        "in": "query",
                        "required": true,
                        "style": "form",
                        "schema": {
                              "type": "string",
                              "example": "123456789aa"
                        }
                  }
            }
      }
}